tinyint tho

+++ 21.09.2023_correct_solutions/03_Bozhidar_Pavlov.sql
@@ -12,7 +12,7 @@ use students;
 create table students(
 	id serial primary key,
 	name varchar(255),
-	age smallint
+	age tinyint
 );
 
 -- (6) Добавяме редове
